Data.MinLen

Type level naturals

data Zero

Constructors

Zero

data Succ nat

Constructors

Succ nat

class TypeNat nat where

Methods

toValueNat :: Num i => nat -> i

typeNat :: nat

type family AddNat x y

type family MaxNat x y

Minimum length newtype wrapper

data MinLen nat mono

unMinLen :: MinLen nat mono -> mono

toMinLenZero :: mono -> MinLen Zero mono

toMinLen :: (MonoFoldable mono, TypeNat nat) => mono -> Maybe (MinLen nat mono)

unsafeToMinLen :: mono -> MinLen nat mono

Although this function itself cannot cause a segfault, it breaks the safety guarantees of MinLen and can lead to a segfault when using otherwise safe functions.

mlcons :: IsSequence seq => Element seq -> MinLen nat seq -> MinLen (Succ nat) seq

mlappend :: IsSequence seq => MinLen x seq -> MinLen y seq -> MinLen (AddNat x y) seq

mlunion :: GrowingAppend mono => MinLen x mono -> MinLen y mono -> MinLen (MaxNat x y) mono

head :: MonoTraversable mono => MinLen (Succ nat) mono -> Element mono

last :: MonoTraversable mono => MinLen (Succ nat) mono -> Element mono

tailML :: IsSequence seq => MinLen (Succ nat) seq -> MinLen nat seq

initML :: IsSequence seq => MinLen (Succ nat) seq -> MinLen nat seq

class (Semigroup mono, MonoFoldable mono) => GrowingAppend mono

olength (x y) >= olength x + olength y

ofoldMap1 :: (MonoFoldable mono, Semigroup m) => (Element mono -> m) -> MinLen (Succ nat) mono -> m

ofold1 :: (MonoFoldable mono, Semigroup (Element mono)) => MinLen (Succ nat) mono -> Element mono

ofoldr1 :: MonoFoldable mono => (Element mono -> Element mono -> Element mono) -> MinLen (Succ nat) mono -> Element mono

ofoldl1' :: MonoFoldable mono => (Element mono -> Element mono -> Element mono) -> MinLen (Succ nat) mono -> Element mono

A variant of 'ofoldl\'' that has no base case, and thus may only be applied to non-empty structures.

'foldl1\'' f = foldl1 f . otoList

maximum :: MonoFoldableOrd mono => MinLen (Succ nat) mono -> Element mono

like Data.List, but not partial on a MonoFoldable

minimum :: MonoFoldableOrd mono => MinLen (Succ nat) mono -> Element mono

like Data.List, but not partial on a MonoFoldable

maximumBy :: MonoFoldable mono => (Element mono -> Element mono -> Ordering) -> MinLen (Succ nat) mono -> Element mono

like Data.List, but not partial on a MonoFoldable

minimumBy :: MonoFoldable mono => (Element mono -> Element mono -> Ordering) -> MinLen (Succ nat) mono -> Element mono

like Data.List, but not partial on a MonoFoldable