Data.MonoTraversable

Description

Type classes mirroring standard typeclasses, but working with monomorphic containers.

The motivation is that some commonly used data types (i.e., ByteString and Text) do not allow for instances of typeclasses like Functor and Foldable, since they are monomorphic structures. This module allows both monomorphic and polymorphic data types to be instances of the same typeclasses.

All of the laws for the polymorphic typeclasses apply to their monomorphic cousins. Thus, even though a MonoFunctor instance for Set could theoretically be defined, it is omitted since it could violate the functor law of omap f . omap g = omap (f . g).

Note that all typeclasses have been prefixed with Mono, and functions have been prefixed with o. The mnemonic for o is "only one," or alternatively "it's mono, but m is overused in Haskell, so we'll use the second letter instead." (Agreed, it's not a great mangling scheme, input is welcome!)

Documentation

type family Element mono

class MonoFunctor mono where

Methods

omap :: (Element mono -> Element mono) -> mono -> mono

class MonoFoldable mono where

Methods

ofoldMap :: Monoid m => (Element mono -> m) -> mono -> m

ofoldr :: (Element mono -> b -> b) -> b -> mono -> b

ofoldl' :: (a -> Element mono -> a) -> a -> mono -> a

otoList :: mono -> [Element mono]

oall :: (Element mono -> Bool) -> mono -> Bool

oany :: (Element mono -> Bool) -> mono -> Bool

onull :: mono -> Bool

olength :: mono -> Int

olength64 :: mono -> Int64

ocompareLength :: Integral i => mono -> i -> Ordering

otraverse_ :: (MonoFoldable mono, Applicative f) => (Element mono -> f b) -> mono -> f ()

ofor_ :: (MonoFoldable mono, Applicative f) => mono -> (Element mono -> f b) -> f ()

omapM_ :: (MonoFoldable mono, Monad m) => (Element mono -> m ()) -> mono -> m ()

oforM_ :: (MonoFoldable mono, Monad m) => mono -> (Element mono -> m ()) -> m ()

ofoldlM :: (MonoFoldable mono, Monad m) => (a -> Element mono -> m a) -> a -> mono -> m a

ofoldMap1Ex :: Semigroup m => (Element mono -> m) -> mono -> m

Note: this is a partial function. On an empty MonoFoldable, it will throw an exception. See Data.NonNull for a total version of this function.

ofoldr1Ex :: (Element mono -> Element mono -> Element mono) -> mono -> Element mono

Note: this is a partial function. On an empty MonoFoldable, it will throw an exception. See Data.NonNull for a total version of this function.

ofoldl1Ex' :: (Element mono -> Element mono -> Element mono) -> mono -> Element mono

Note: this is a partial function. On an empty MonoFoldable, it will throw an exception. See Data.NonNull for a total version of this function.

headEx :: mono -> Element mono

lastEx :: mono -> Element mono

unsafeHead :: mono -> Element mono

unsafeLast :: mono -> Element mono

maximumByEx :: (Element mono -> Element mono -> Ordering) -> mono -> Element mono

minimumByEx :: (Element mono -> Element mono -> Ordering) -> mono -> Element mono

headMay :: MonoFoldable mono => mono -> Maybe (Element mono)

like Data.List.head, but not partial

lastMay :: MonoFoldable mono => mono -> Maybe (Element mono)

like Data.List.last, but not partial

osum :: (MonoFoldable mono, Num (Element mono)) => mono -> Element mono

The sum function computes the sum of the numbers of a structure.

oproduct :: (MonoFoldable mono, Num (Element mono)) => mono -> Element mono

The product function computes the product of the numbers of a structure.

oand :: (Element mono ~ Bool, MonoFoldable mono) => mono -> Bool

Are all of the values True?

Since 0.6.0

oor :: (Element mono ~ Bool, MonoFoldable mono) => mono -> Bool

Are any of the values True?

Since 0.6.0

class (MonoFoldable mono, Monoid mono) => MonoFoldableMonoid mono where

Methods

oconcatMap :: (Element mono -> mono) -> mono -> mono

class (MonoFoldable mono, Ord (Element mono)) => MonoFoldableOrd mono where

A typeclass for MonoFoldables containing elements which are an instance of Ord.

Methods

maximumEx :: mono -> Element mono

minimumEx :: mono -> Element mono

maximumMay :: MonoFoldableOrd mono => mono -> Maybe (Element mono)

maximumByMay :: MonoFoldable mono => (Element mono -> Element mono -> Ordering) -> mono -> Maybe (Element mono)

minimumMay :: MonoFoldableOrd mono => mono -> Maybe (Element mono)

minimumByMay :: MonoFoldable mono => (Element mono -> Element mono -> Ordering) -> mono -> Maybe (Element mono)

class (MonoFunctor mono, MonoFoldable mono) => MonoTraversable mono where

Methods

otraverse :: Applicative f => (Element mono -> f (Element mono)) -> mono -> f mono

omapM :: Monad m => (Element mono -> m (Element mono)) -> mono -> m mono

ofor :: (MonoTraversable mono, Applicative f) => mono -> (Element mono -> f (Element mono)) -> f mono

oforM :: (MonoTraversable mono, Monad f) => mono -> (Element mono -> f (Element mono)) -> f mono

ofoldlUnwrap :: MonoFoldable mono => (x -> Element mono -> x) -> x -> (x -> b) -> mono -> b

A strict left fold, together with an unwrap function.

This is convenient when the accumulator value is not the same as the final expected type. It is provided mainly for integration with the foldl package, to be used in conjunction with purely.

Since 0.3.1

ofoldMUnwrap :: (Monad m, MonoFoldable mono) => (x -> Element mono -> m x) -> m x -> (x -> m b) -> mono -> m b

A monadic strict left fold, together with an unwrap function.

Similar to foldlUnwrap, but allows monadic actions. To be used with impurely from foldl.

Since 0.3.1

class MonoPointed mono where

Instances must obey the laws:

otoList . mconcat . map opoint == id

Methods

opoint :: Element mono -> mono